requirements.txt
numpy>=1.24.0,<2.0
Pillow>=10.0.0
PyYAML>=6.0.0
tqdm>=4.66.0
# Optional for simplified 3D prior backend
# pytorch3d>=0.7.7
